+++ cmd/hlc/main.go
@@ -16,8 +16,22 @@ import (
 func main() {
 	args := os.Args[1:]
 
-	if flag.GetBoolFlagValue(args, "version") {
+	if flag.GetBoolFlagValue(args, []string{"v", "version"}) {
 		fmt.Println(hiddenlake.GVersion)
+		return
+	}
+
+	if flag.GetBoolFlagValue(args, []string{"h", "help"}) {
+		fmt.Print(
+			"Hidden Lake Composite (HLC)\n" +
+				"Description: runs many HL services\n" +
+				"Arguments:\n" +
+				"[ -h, --help    ] - print information about service\n" +
+				"[ -v, --version ] - print version of service\n" +
+				"[ -p, --path    ] - set path to config, database files\n" +
+				"[ -n, --network ] - set network key for connections\n" +
+				"[ -t, --threads ] - set num of parallel functions to calculate PoW\n",
+		)
 		return
 	}