+++ internal/applications/messenger/pkg/app/init_app.go
@@ -12,7 +12,7 @@ import (
 )
 
 func InitApp(pArgs []string) (types.IRunner, error) {
-	inputPath := strings.TrimSuffix(flag.GetFlagValue(pArgs, "path", "."), "/")
+	inputPath := strings.TrimSuffix(flag.GetFlagValue(pArgs, []string{"p", "path"}, "."), "/")
 
 	cfgPath := filepath.Join(inputPath, settings.CPathYML)
 	cfg, err := config.InitConfig(cfgPath, nil)