+++ internal/utils/flag/flag_test.go
@@ -14,7 +14,7 @@ func TestPanicFlagValue(t *testing.T) {
 	argsSlice := []string{
 		"--key",
 	}
-	_ = GetFlagValue(argsSlice, "key", "_")
+	_ = GetFlagValue(argsSlice, []string{"key"}, "_")
 }
 
 func TestBoolFlagValue(t *testing.T) {
@@ -26,32 +26,32 @@ func TestBoolFlagValue(t *testing.T) {
 		"value", "571",
 	}
 
-	if !GetBoolFlagValue(argsSlice, "key") {
+	if !GetBoolFlagValue(argsSlice, []string{"key"}) {
 		t.Error("!key")
 		return
 	}
 
-	if !GetBoolFlagValue(argsSlice, "123") {
+	if !GetBoolFlagValue(argsSlice, []string{"123"}) {
 		t.Error("!123")
 		return
 	}
 
-	if !GetBoolFlagValue(argsSlice, "name") {
+	if !GetBoolFlagValue(argsSlice, []string{"name"}) {
 		t.Error("!name")
 		return
 	}
 
-	if !GetBoolFlagValue(argsSlice, "value") {
+	if !GetBoolFlagValue(argsSlice, []string{"value"}) {
 		t.Error("!value")
 		return
 	}
 
-	if !GetBoolFlagValue(argsSlice, "571") {
+	if !GetBoolFlagValue(argsSlice, []string{"571"}) {
 		t.Error("!571")
 		return
 	}
 
-	if GetBoolFlagValue(argsSlice, "undefined") {
+	if GetBoolFlagValue(argsSlice, []string{"undefined"}) {
 		t.Error("success get undefined value")
 		return
 	}
@@ -70,37 +70,37 @@ func TestFlagValue(t *testing.T) {
 		"--zxcvb=!@#$%",
 	}
 
-	if GetFlagValue(argsSlice, "key", "1") != "123" {
+	if GetFlagValue(argsSlice, []string{"key"}, "1") != "123" {
 		t.Error("key != 123")
 		return
 	}
 
-	if GetFlagValue(argsSlice, "name", "2") != "number" {
+	if GetFlagValue(argsSlice, []string{"name"}, "2") != "number" {
 		t.Error("name != number")
 		return
 	}
 
-	if GetFlagValue(argsSlice, "value", "3") != "571" {
+	if GetFlagValue(argsSlice, []string{"value"}, "3") != "571" {
 		t.Error("value != 571")
 		return
 	}
 
-	if GetFlagValue(argsSlice, "asdfg", "4") != "12345" {
+	if GetFlagValue(argsSlice, []string{"asdfg"}, "4") != "12345" {
 		t.Error("asdfg != 12345")
 		return
 	}
 
-	if GetFlagValue(argsSlice, "qwerty", "5") != "67890" {
+	if GetFlagValue(argsSlice, []string{"qwerty"}, "5") != "67890" {
 		t.Error("qwerty != 67890")
 		return
 	}
 
-	if GetFlagValue(argsSlice, "zxcvb", "6") != "!@#$%" {
+	if GetFlagValue(argsSlice, []string{"zxcvb"}, "6") != "!@#$%" {
 		t.Error("zxcvb != !@#$%")
 		return
 	}
 
-	if GetFlagValue(argsSlice, "unknown", "7") != "7" {
+	if GetFlagValue(argsSlice, []string{"unknown"}, "7") != "7" {
 		t.Error("unknown != 7")
 		return
 	}